+++ src/v1/BasicTypes/HashMaps/HashMap.ts
@@ -32,6 +32,9 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
+import { findAttributeNames } from "../Objects";
+import { STOP_AT_NEXT_PROTOTYPE } from "../Prototypes";
+
 /**
  * `HashMap` describes an object that doesn't have a set list of keys.
  *
@@ -42,4 +45,29 @@
  */
 export interface HashMap<T> {
     [key: string]: T
+}
+
+export class HashMap<T> {
+    /**
+     * `forEach()` is the equivalent of {@link Array.forEach}, only
+     * it works on {@link HashMap}s instead.
+     *
+     * @param target
+     * the HashMap to iterate over
+     * @param callbackfn
+     * the function to call when we iterate
+     *
+     * @category BasicTypes
+     */
+    public static forEach<T>(
+        target: HashMap<T>,
+        callbackfn: (value: T, name: string, obj: HashMap<T>) => void
+    ): void {
+        findAttributeNames(
+            target,
+            { nextPrototype: STOP_AT_NEXT_PROTOTYPE }
+        ).forEach((name: string) => {
+            callbackfn(target[name], name, target);
+        });
+    }
 }