+++ CHANGELOG.md
@@ -21,6 +21,23 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v0.2.1
+
+Released Sunday, 6th September 2020.
+
+### New
+
+* Added `HashMap.forEach()`
+
+### Tests
+
+* `extractReasonFromCaught()` and `extractStackFromCaught()`
+  - unit tests no longer look for `Context.<anonymous>` in the stack trace
+
+### Dependencies
+
+* Updated dev dependencies to resolve low-severity `npm audit` warnings.
+
 ## v0.2.0
 
 Released Friday, 10th July 2020.